requirements.txt
aioredis==2.0.1
anyio==3.7.1
async-timeout==4.0.3
cffi==1.16.0
cleo==1.0.0a4
click==8.1.7
crashtest==0.3.1
cryptography==42.0.0
exceptiongroup==1.2.0
fastapi==0.109.0
hiredis==2.3.2
idna==3.6
more-itertools==9.1.0
pptree==3.1
pycparser==2.21
pydantic==1.10.14
pylev==1.4.0
python-dotenv==0.19.2
python-ulid==1.1.0
redis==4.6.0
redis-om==0.2.1
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
types-pyOpenSSL==23.3.0.20240106
types-redis==4.6.0.20240106
types-six==1.16.21.20240106
typing-extensions==4.9.0
uvloop==0.19.0
